float32

+++ models/official/resnet/imagenet_input.py
@@ -231,10 +231,10 @@ class ImageNetTFExampleInput(object):
     
     if not self.use_bfloat16:
       prices = tf.cast(prices, tf.float32)
-      label = tf.cast(label, tf.int64)
+      label = tf.cast(label, tf.float32)
     else:
       prices = tf.cast(prices, tf.bfloat16)
-      label = tf.cast(label, tf.int64)
+      label = tf.cast(label, tf.float32)
     
     
     #prices = tf.reshape(prices, [-1])
